Правки после второго ревью
